move StringMatcher to package stringtool
