+++ pkg/filters/proxies/requestmatch_test.go
@@ -25,7 +25,6 @@ import (
 	"testing"
 
 	"github.com/megaease/easegress/pkg/protocols/httpprot"
-	"github.com/stretchr/testify/assert"
 )
 
 func TestRandomMatcher(t *testing.T) {
@@ -103,47 +102,4 @@ func TestIPHashMatcher(t *testing.T) {
 	if match < 900 || match > 1100 {
 		t.Errorf("ip hash matcher is not working as configured")
 	}
-}
-
-func TestStringMatcher(t *testing.T) {
-	assert := assert.New(t)
-
-	// validation
-	sm := &StringMatcher{Empty: true}
-	assert.NoError(sm.Validate())
-	sm.Init()
-
-	sm = &StringMatcher{Empty: true, Exact: "abc"}
-	assert.Error(sm.Validate())
-
-	sm = &StringMatcher{}
-	assert.Error(sm.Validate())
-
-	sm = &StringMatcher{RegEx: "^abc[0-9]+$"}
-	assert.NoError(sm.Validate())
-	sm.Init()
-
-	sm.Prefix = "/xyz"
-	assert.NoError(sm.Validate())
-
-	sm.Exact = "/abc"
-	assert.NoError(sm.Validate())
-
-	// match
-	sm = &StringMatcher{Empty: true}
-	assert.True(sm.Match(""))
-	assert.False(sm.Match("abc"))
-
-	sm = &StringMatcher{RegEx: "^abc[0-9]+$"}
-	sm.Init()
-	assert.True(sm.Match("abc123"))
-	assert.False(sm.Match("abc123d"))
-
-	sm.Prefix = "/xyz"
-	assert.True(sm.Match("/xyz123"))
-	assert.False(sm.Match("/Xyz123"))
-
-	sm.Exact = "/hello"
-	assert.True(sm.Match("/hello"))
-	assert.False(sm.Match("/Hello"))
 }
